+++ Project_3/include/SimFunctions.h
@@ -15,14 +15,14 @@ class SimFunctions
     public:
         SimFunctions();
         virtual ~SimFunctions();
-        virtual float fallFactor(Pitch& p);
-        virtual float baseTension(Pitch& p);
+        virtual double fallFactor(Pitch& p);
+        virtual double baseTension(Pitch& p);
        // virtual Matrix incrementalSlips(Pitch &p);
-        virtual float delS(float t, float v);
+        virtual double delS(double t, double v);
     protected:
     private:
        // Pitch p; //have a pitch in the thing??
-       float tInc; //how much to increase time at each time step
+       double tInc; //how much to increase time at each time step
 
 };
